ICA atlas

from Abstract: "We present an integrated approach to probabilistic independent component analysis (ICA) for functional MRI (FMRI) data that allows for nonsquare mixing in the presence of Gaussian noise. In order to avoid overfitting, we employ objective estimation of the amount of Gaussian noise through Bayesian analysis of the true dimensionality of the data, i.e., the number of activation and non-Gaussian noise sources. This enables us to carry out probabilistic modeling and achieves an asymptotically unique decomposition of the data. It reduces problems of interpretation, as each final independent component is now much more likely to be due to only one physical or physiological process.

ica

ica_3d

Format

An object of class brain_atlas of length 4.

An object of class ggseg3d_atlas (inherits from tbl_df, tbl, data.frame) with 4 rows and 4 columns.

References

Beckmann, C. F., & Smith, S. M. (2004). Probabilistic independent component analysis for functional magnetic resonance imaging. IEEE transactions on medical imaging, 23(2), 137-152. (IEEEE)

ica - ica atlas

ica_3d - ica 3d mesh atlas

Examples

data(ica)
data(ica_3d)